ASoC: fsl_ssi: Configure the RX/TX if it's unconfigured
In case the RX/TX is unconfigured, the SDMA won't be able to push
the audio stream to the codec. It is normal that the nr_active_streams
is non-zero upon first invocation of fsl_ssi_config() in AC97 mode,
but this does prevent the configuration of the RX/TX from happening.
Instead of depending on the number of active streams being zero when
starting the configuration, check whether the RX/TX was configured
already and if not, configure it.

Signed-off-by: Marek Vasut <marex@denx.de>