PCI: imx6sx: Speed up the resume time considerably
In 886bc5c, a generic loop to wait for the link was introduced.
If the link wasn't already up, this increased the wait time to at least 100ms
until the next check. This patch increases the number of checks and decreases
the wait time for the link check.

In addition, 100ms between toggling the PERST line seems rather long and
isn't necessary for the FPGA to function property.